requirements.txt
absl-py==2.1.0
# For V-DROID Training
torch
transformers
datasets
peft
trl
bitsandbytes
accelerate

android_env==1.2.3
dm_env==1.6
easyocr
fuzzywuzzy==0.18.0
google-generativeai
grpcio-tools
immutabledict==2.0.0
IPython
jsonschema==4.17.3
matplotlib==3.6.1
numpy==1.26.3
ollama
opencv-python
pandas==2.1.4
Pillow
pydub
python-Levenshtein
pytest
pytesseract
requests
tenacity
termcolor